requirements.txt
Flask~=1.1.2
google-cloud-pubsub
google~=3.0.0
google-cloud-storage
google-cloud-secret-manager
structlog==20.1.0
python-gnupg==0.4.6